Add files via upload
This is an end-to-end program for sentiment Analysis. I have trained the LSTM model to understand the sentiments of the reviewers. The data that I have used In this project is the amazon reviews dataset that is available for free. I have used Flask to deploy the model on my local system. This project contains all the files that you will require to run this at your end. Put the Flask_app folder inside your PyCharm environment and run the app_Sentiment_Analysis.py file.  Go to http://127.0.0.1:5000/ on your browser. Write the Review and test the model for yourself.